+++ interface.go
@@ -1,8 +1,8 @@
 package gnames
 
 import (
-	"github.com/gnames/gnlib/domain/entity/gn"
-	vlib "github.com/gnames/gnlib/domain/entity/verifier"
+	"github.com/gnames/gnlib/ent/gnvers"
+	vlib "github.com/gnames/gnlib/ent/verifier"
 )
 
 // GNames is the main use-case interface of the app. Its purpose to provide
@@ -11,7 +11,12 @@ import (
 // known to gnames scientific names, as well as providing data where
 // these names occur.
 type GNames interface {
-	gn.Versioner
+	// GetVersion returns the version of GNames and a timestamp of its build.
+	GetVersion() gnvers.Version
+	// Verify takes a slice of name-strings together with query parameters and
+	// returns back results of verification.
 	Verify(params vlib.VerifyParams) ([]*vlib.Verification, error)
+	// Datasources take IDs of data-sourses and return back list of corresponding
+	// metadata. If no IDs are given, it returns metadata for all data-sources.
 	DataSources(ids ...int) ([]*vlib.DataSource, error)
 }