+++ io/matcher/matcher_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"github.com/gnames/gnames/io/matcher"
-	vlib "github.com/gnames/gnlib/domain/entity/verifier"
+	vlib "github.com/gnames/gnlib/ent/verifier"
 	"github.com/gnames/gnmatcher"
 	"github.com/stretchr/testify/assert"
 )
@@ -12,15 +12,15 @@ import (
 const url = "https://matcher.globalnames.org/api/v1/"
 
 func TestVer(t *testing.T) {
-	var m gnmatcher.GNMatcher
-	m = matcher.NewGNMatcher(url)
+	var m gnmatcher.GNmatcher
+	m = matcher.NewGNmatcher(url)
 	ver := m.GetVersion()
 	assert.Regexp(t, `^v\d+\.\d+\.\d+`, ver.Version)
 }
 
 func TestMatch(t *testing.T) {
-	var m gnmatcher.GNMatcher
-	m = matcher.NewGNMatcher(url)
+	var m gnmatcher.GNmatcher
+	m = matcher.NewGNmatcher(url)
 	res := m.MatchNames([]string{"Pardosa moeste"})
 	assert.Equal(t, res[0].Name, "Pardosa moeste")
 	assert.Equal(t, res[0].MatchType, vlib.Fuzzy)