+++ io/rest/interface.go
@@ -2,8 +2,7 @@ package rest
 
 import (
 	"github.com/gnames/gnames"
-	"github.com/gnames/gnlib/domain/entity/gn"
-	"github.com/gnames/gnlib/encode"
+	"github.com/gnames/gnfmt"
 )
 
 // VerifierService interface is the API behing RESTful service.
@@ -11,13 +10,10 @@ type VerifierService interface {
 	// Ping checks if the service is alive.
 	Ping() string
 
-	// Versioner returns Version of gnames project.
-	gn.Versioner
-
 	// Port returns port of the HTTP/1 service.
 	Port() int
 
 	gnames.GNames
 
-	encode.Encoder
+	gnfmt.Encoder
 }